fix(typetest): rename files as per convention
